feat(openapi): manage error resources in global
