+++ src/Metadata/Operation.php
@@ -13,6 +13,7 @@ declare(strict_types=1);
 
 namespace ApiPlatform\Metadata;
 
+use ApiPlatform\Metadata\Exception\ProblemExceptionInterface;
 use ApiPlatform\State\OptionsInterface;
 
 /**
@@ -47,19 +48,20 @@ abstract class Operation extends Metadata
      *     class?: string|null,
      *     name?: string,
      * }|string|false|null $output {@see https://api-platform.com/docs/core/dto/#specifying-an-input-or-an-output-data-representation}
-     * @param string|array|bool|null   $mercure       {@see https://api-platform.com/docs/core/mercure}
-     * @param string|bool|null         $messenger     {@see https://api-platform.com/docs/core/messenger/#dispatching-a-resource-through-the-message-bus}
-     * @param bool|null                $elasticsearch {@see https://api-platform.com/docs/core/elasticsearch/}
-     * @param bool|null                $read          {@see https://api-platform.com/docs/core/events/#the-event-system}
-     * @param bool|null                $deserialize   {@see https://api-platform.com/docs/core/events/#the-event-system}
-     * @param bool|null                $validate      {@see https://api-platform.com/docs/core/events/#the-event-system}
-     * @param bool|null                $write         {@see https://api-platform.com/docs/core/events/#the-event-system}
-     * @param bool|null                $serialize     {@see https://api-platform.com/docs/core/events/#the-event-system}
-     * @param bool|null                $fetchPartial  {@see https://api-platform.com/docs/core/performance/#fetch-partial}
-     * @param bool|null                $forceEager    {@see https://api-platform.com/docs/core/performance/#force-eager}
-     * @param string|callable|null     $provider      {@see https://api-platform.com/docs/core/state-providers/#state-providers}
-     * @param string|callable|null     $processor     {@see https://api-platform.com/docs/core/state-processors/#state-processors}
-     * @param array<string, Parameter> $parameters
+     * @param string|array|bool|null                                $mercure       {@see https://api-platform.com/docs/core/mercure}
+     * @param string|bool|null                                      $messenger     {@see https://api-platform.com/docs/core/messenger/#dispatching-a-resource-through-the-message-bus}
+     * @param bool|null                                             $elasticsearch {@see https://api-platform.com/docs/core/elasticsearch/}
+     * @param bool|null                                             $read          {@see https://api-platform.com/docs/core/events/#the-event-system}
+     * @param bool|null                                             $deserialize   {@see https://api-platform.com/docs/core/events/#the-event-system}
+     * @param bool|null                                             $validate      {@see https://api-platform.com/docs/core/events/#the-event-system}
+     * @param bool|null                                             $write         {@see https://api-platform.com/docs/core/events/#the-event-system}
+     * @param bool|null                                             $serialize     {@see https://api-platform.com/docs/core/events/#the-event-system}
+     * @param bool|null                                             $fetchPartial  {@see https://api-platform.com/docs/core/performance/#fetch-partial}
+     * @param bool|null                                             $forceEager    {@see https://api-platform.com/docs/core/performance/#force-eager}
+     * @param string|callable|null                                  $provider      {@see https://api-platform.com/docs/core/state-providers/#state-providers}
+     * @param string|callable|null                                  $processor     {@see https://api-platform.com/docs/core/state-processors/#state-processors}
+     * @param array<string, Parameter>                              $parameters
+     * @param array<class-string<ProblemExceptionInterface>>|null   $errors
      */
     public function __construct(
         protected ?string $shortName = null,
@@ -806,6 +808,7 @@ abstract class Operation extends Metadata
         protected $provider = null,
         protected $processor = null,
         protected ?OptionsInterface $stateOptions = null,
+        ?array $errors = null,
         array|Parameters|null $parameters = null,
         array|string|null $rules = null,
         ?string $policy = null,
@@ -853,6 +856,7 @@ abstract class Operation extends Metadata
             provider: $provider,
             processor: $processor,
             stateOptions: $stateOptions,
+            errors: $errors,
             parameters: $parameters,
             rules: $rules,
             policy: $policy,
